Added "DataSummary": moved events to this class.
Realized that UsageEvent can (and should) also be used as INPUT to a EWH
simulation.

Renamed "GeyserData" to "DataSegment". Each class name does not have to
specify its "Geyser" affinity since it is under the "geyserM2M"
namespace. (Remember: Namespace is not just a technicality to avoid
conflicts with the rest of the world. It is also a logical way of
thinking about classes and their structure to the whole)